pyproject.toml
[tool.poetry]
name = "ai-website-generator"
version = "0.1.0"
description = "AI-powered website builder using CrewAI for simple React applications"
authors = ["AI Website Builder Team"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = "^0.121.1"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
anthropic = "^0.7.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
aiofiles = "^23.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.7.0"
isort = "^5.12.0"

[tool.poetry.group.test.dependencies]
pytest-cov = "^4.1.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
